🚚 Move error componenets to page components
