+++ src/lib/PageComponents/Error/AuthError.svelte
